+++ app/features/auth/components/index.ts
@@ -1,2 +1,4 @@
+export { LoginModal } from "./login-modal";
+export { SignoutButton } from "./signout-button";
 export { WithrawButton } from "./withraw-button";
 export { WithrawDialog } from "./withraw-dialog";
