--- app/features/auth/login/components/index.ts
@@ -1,1 +1,0 @@
-export { LoginModal } from "./login-modal";
