+++ app/features/user/components/index.ts
@@ -1,4 +1,3 @@
-export { SignoutButton } from "./signout-button";
 export { CategoryTabs } from "./category-tabs";
 export { MyReviewCard } from "./my-review-card";
 export { MyReviews } from "./my-review-list";
